HUGE backend API refactor, no frontend connection /url changes but cleaned up code. split off /videos into its own sub app file
